requirements.txt
# Core dependencies
psutil>=5.9.0
structlog>=23.1.0
pyyaml>=6.0.1
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0

# eBPF dependencies
# bcc>=0.28.0

# Monitoring dependencies
prometheus-client>=0.16.0

# Test dependencies
pytest>=7.3.1
pytest-cov>=4.1.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
pytest-timeout>=2.1.0
pytest-xdist>=3.3.1
coverage>=7.2.5
